Pull SearchDatasetTutorial steps to config
